🧹 cleanup: Extract duplicated tree resolution logic
- Extracted `TreeContext` and tree resolution logic into `TreeContextResolver` utility.
- Updated `StickyScrollManager` and `PinnedFoldersManager` to use the shared utility.
- Improved maintainability by removing copy-pasted code.
