+++ src/main/kotlin/com/github/erjotek/stickyprojectfolder/ui/StickyScrollManager.kt
@@ -10,6 +10,7 @@ import com.intellij.openapi.project.Project
 import com.intellij.openapi.wm.ToolWindowId
 import com.intellij.openapi.wm.ToolWindowManager
 import com.intellij.openapi.wm.ex.ToolWindowManagerListener
+import com.github.erjotek.stickyprojectfolder.util.TreeContextResolver
 import com.intellij.ui.content.ContentManagerEvent
 import com.intellij.ui.content.ContentManagerListener
 import java.awt.*
@@ -96,7 +97,7 @@ class StickyScrollManager(private val project: Project) : Disposable {
             toolWindow.contentManager.addContentManagerListener(listener)
         }
 
-        val context = resolveTreeContext(pane)
+        val context = TreeContextResolver.resolve(project, pane)
         if (context == null) {
             if (this.tree != null) {
                 detach()
@@ -159,53 +160,6 @@ class StickyScrollManager(private val project: Project) : Disposable {
             com.intellij.openapi.util.Disposer.register(this, autoCollapseManager!!)
             autoCollapseManager?.install()
         }
-    }
-
-    private data class TreeContext(val tree: JTree, val scrollPane: JScrollPane)
-
-    private fun resolveTreeContext(pane: AbstractProjectViewPane?): TreeContext? {
-        fun toContext(tree: JTree): TreeContext? {
-            val sp = SwingUtilities.getAncestorOfClass(JScrollPane::class.java, tree) as? JScrollPane ?: return null
-            return TreeContext(tree, sp)
-        }
-
-        resolveFocusedProjectViewTree()?.let { focusedTree ->
-            toContext(focusedTree)?.let { return it }
-        }
-
-        val treeFromPane = pane?.tree
-        if (treeFromPane != null && treeFromPane.isShowing) {
-            toContext(treeFromPane)?.let { return it }
-        }
-
-        val toolWindow = ToolWindowManager.getInstance(project).getToolWindow(ToolWindowId.PROJECT_VIEW) ?: return null
-        val treeFromToolWindow = findTree(toolWindow.component) ?: return null
-        return toContext(treeFromToolWindow)
-    }
-
-    private fun resolveFocusedProjectViewTree(): JTree? {
-        val toolWindow = ToolWindowManager.getInstance(project).getToolWindow(ToolWindowId.PROJECT_VIEW) ?: return null
-        val toolWindowComponent = toolWindow.component
-
-        val focusOwner = KeyboardFocusManager.getCurrentKeyboardFocusManager().focusOwner ?: return null
-        if (!SwingUtilities.isDescendingFrom(focusOwner, toolWindowComponent)) return null
-
-        return when (focusOwner) {
-            is JTree -> focusOwner
-            else -> SwingUtilities.getAncestorOfClass(JTree::class.java, focusOwner) as? JTree
-        }
-    }
-
-    private fun findTree(component: Component?): JTree? {
-        if (component == null) return null
-        if (component is JTree && component.isShowing) return component
-        if (component is Container) {
-            for (child in component.components) {
-                val tree = findTree(child)
-                if (tree != null) return tree
-            }
-        }
-        return null
     }
 
     private fun installListeners(sp: JScrollPane, currentTree: JTree) {
